Time: 2 ms (26.62%) | Memory: 62 MB (78.42%) - LeetSync

+++ 4005-maximum-total-subarray-value-i/maximum-total-subarray-value-i.java
@@ -8,6 +8,6 @@ class Solution {
             max = Math.max(num,max);
             min = Math.min(num,min);
         }
-        return (max-min)*k;
+        return (long)(max-min)*k;
     }
 }
